int function for the "how_many" variable in line 24
I have tried the original code on Python 3.8.2 but it kept showing errors " 'str' object cannot be interpreted as an integer". While reading the traceback, I noticed in line 24, the variable "how_many" was acted as a string variable, not an integer one (I would love if someone explained to me why even though it was defined as an integer the function's argument). The only thing I did was to add int() to the "how_many" variable to convert it into an integer one. When I tested it back on my device, it worked like charm